requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
colorama==0.4.4
Faker==8.12.1
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.2.3
Flask-Migrate==3.0.1
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.21
SQLAlchemy-Utils==0.37.8
text-unidecode==1.3
typing-extensions==3.10.0.0
Werkzeug==0.16.1
zipp==3.5.0
